renamed files, serviceworker registered
