+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import Header from './components/header'
-import Footer from './components/footer'
-import Home from './app/home'
+import Header from './components/header/header'
+import Footer from './components/footer/footer'
+import Home from './app/home/home'
 import './App.scss';
 import './assets/styles/flex.scss'
